requirements.txt
pyTelegramBotAPI
feedparser
gTTS
requests
openai
python-dotenv
